refactor(env): update import paths for env config
